+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 ## [Unreleased]
 
 
+## [0.24.6] - 2019-03-19
+### Changed
+- [RPC] 'blocks' on info RPC now lists blocks in ascending rather than descending height order
+
+### Added
+- [CLI] Introduced burrow configure --pool for generation of multiple validator configs suitable for running on a single (or many) machines
+
+### Fixed
+- [Metrics] Fix histogram statistics by making counts cumulative 
+
+
 ## [0.24.5] - 2019-03-14
 ### Changed
 - [Consensus] Tendermint timeouts configurable by a float factor from defaults and default change to 0.33 of Tendermint's default for smaller networks'
@@ -245,7 +256,7 @@ In addition to breaking changes associated with Tendermint (see their changelog)
 - Add BlockExplorer CLI for forensics
 - Expose reason for REVERT
 - Add last_block_info healthcheck endpoint to rpc/TM
-- 
+-
 #### Improvements
 - Implement checkpointing when saving application and blockchain state in commit - interrupted commit rolls burrow back to last block whereon it can catch up using Tendermint
 - Maintain separate read-only tree in state so that long-running RPC request cannot block writes
@@ -411,7 +422,7 @@ This release marks the start of Eris-DB as the full permissioned blockchain node
   - Allow multiple event subscriptions from same host under rpc/tendermint
 
 
-- Tool changes  
+- Tool changes
   - Use glide instead of godeps for dependencies
 
 
@@ -428,7 +439,8 @@ This release marks the start of Eris-DB as the full permissioned blockchain node
   - [Blockchain] Fix getBlocks to respect block height cap.
 
 
-[Unreleased]: https://github.com/hyperledger/burrow/compare/v0.24.5...HEAD
+[Unreleased]: https://github.com/hyperledger/burrow/compare/v0.24.6...HEAD
+[0.24.6]: https://github.com/hyperledger/burrow/compare/v0.24.5...v0.24.6
 [0.24.5]: https://github.com/hyperledger/burrow/compare/v0.24.4...v0.24.5
 [0.24.4]: https://github.com/hyperledger/burrow/compare/v0.24.3...v0.24.4
 [0.24.3]: https://github.com/hyperledger/burrow/compare/v0.24.2...v0.24.3